section_4/5_encryption/requirements.txt
pycryptodome==3.4.7